Refactor data type
