+++ wormhole-infra/wormhole-context/src/main/java/com/zergclan/wormhole/creator/TaskMetadataCreator.java
@@ -23,13 +23,13 @@ import com.zergclan.wormhole.core.config.SourceConfiguration;
 import com.zergclan.wormhole.core.config.TargetConfiguration;
 import com.zergclan.wormhole.core.config.TaskConfiguration;
 import com.zergclan.wormhole.core.metadata.DataSourceMetadata;
-import com.zergclan.wormhole.core.metadata.plan.DataNodeMetadata;
-import com.zergclan.wormhole.core.metadata.plan.FilterMetadata;
-import com.zergclan.wormhole.core.metadata.plan.LinkType;
-import com.zergclan.wormhole.core.metadata.plan.LoadType;
-import com.zergclan.wormhole.core.metadata.plan.SourceMetadata;
-import com.zergclan.wormhole.core.metadata.plan.TargetMetadata;
-import com.zergclan.wormhole.core.metadata.plan.TaskMetadata;
+import com.zergclan.wormhole.core.metadata.node.DataNodeMetadata;
+import com.zergclan.wormhole.core.metadata.task.FilterMetadata;
+import com.zergclan.wormhole.core.metadata.task.LinkType;
+import com.zergclan.wormhole.core.metadata.task.LoadType;
+import com.zergclan.wormhole.core.metadata.task.SourceMetadata;
+import com.zergclan.wormhole.core.metadata.task.TargetMetadata;
+import com.zergclan.wormhole.core.metadata.task.TaskMetadata;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
