+++ wormhole-infra/wormhole-core/src/main/java/com/zergclan/wormhole/core/metadata/plan/PlanMetadata.java
@@ -19,6 +19,7 @@ package com.zergclan.wormhole.core.metadata.plan;
 
 import com.zergclan.wormhole.core.metadata.Metadata;
 import com.zergclan.wormhole.core.metadata.Refreshable;
+import com.zergclan.wormhole.core.metadata.task.TaskMetadata;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
