+++ wormhole-infra/wormhole-core/src/main/java/com/zergclan/wormhole/core/metadata/resource/ColumnMetadata.java
@@ -23,7 +23,7 @@ import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 /**
- * Metadata for column.
+ * Column metadata.
  */
 @RequiredArgsConstructor
 @Getter
