+++ wormhole-infra/wormhole-core/src/main/java/com/zergclan/wormhole/core/metadata/resource/SchemaMetadata.java
@@ -26,7 +26,7 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 
 /**
- * Metadata of schema.
+ * Schema metadata.
  */
 @RequiredArgsConstructor
 @Getter
